Reorganize dir.targets files
The samples, src, and tools directories now have their
own dir.targets files, instead of them all using the
same file at the root of the repository.

This change also emphasizes the fact that the dir.targets
files are private implementation details of the repository
itself, and are not intended to be imported by outside projects.