Add Microsoft.Win32.Resources project
This project is a copy of Vestris.ResourceLib with a different namespace. I cannot use the
version from NuGet because it does not list support for .NET Core in its nuspec file.

Note that I am placing this assembly into src/ instead of tools/ because library code
will also reference it.