Remove McgIntrinsics class
This code was added to support CoreRT in Win32UI, a development angle that has
since been abandoned for now. Additionally, if I ever decide to go back to CoreRT,
chances are that this could would no longer be needed anyways. (I added this
class to workaround the lack of a working Marshal.GetFunctionPointerForDelegate()
method under CoreRT, as had existed at the time.)